Remove duplication around _update_from_file.
