+++ lib/galaxy/objectstore/azure_blob.py
@@ -313,34 +313,6 @@ class AzureBlobObjectStore(ConcreteObjectStore, UsesCache):
             log.exception("%s delete error", self._get_filename(obj, **kwargs))
         return False
 
-    def _update_from_file(self, obj, file_name=None, create=False, **kwargs):
-        if create is True:
-            self._create(obj, **kwargs)
-
-        if self._exists(obj, **kwargs):
-            rel_path = self._construct_path(obj, **kwargs)
-            # Chose whether to use the dataset file itself or an alternate file
-            if file_name:
-                source_file = os.path.abspath(file_name)
-                # Copy into cache
-                cache_file = self._get_cache_path(rel_path)
-                try:
-                    if source_file != cache_file and self.cache_updated_data:
-                        # FIXME? Should this be a `move`?
-                        shutil.copy2(source_file, cache_file)
-                    fix_permissions(self.config, cache_file)
-                except OSError:
-                    log.exception("Trouble copying source file '%s' to cache '%s'", source_file, cache_file)
-            else:
-                source_file = self._get_cache_path(rel_path)
-
-            self._push_to_os(rel_path, source_file)
-
-        else:
-            raise ObjectNotFound(
-                f"objectstore.update_from_file, object does not exist: {str(obj)}, kwargs: {str(kwargs)}"
-            )
-
     def _get_object_url(self, obj, **kwargs):
         if self._exists(obj, **kwargs):
             rel_path = self._construct_path(obj, **kwargs)
